# feat(dht): Optimize `PeerManager#getClosestContactsTo` and `PeerManager#getClosestNeighborsTo`