Optimized the methods by defining `maxSize` for `SortedContactList` instances.

## Benchmarks

- 200 contacts in the `PeerManager`
- find 5 closest  nodes
- 2 random nodes excluded
- default settings (`maxNeighborListSize`: 200, `numberOfNodesPerKBucket`: 8)
- do 1000 queries

`getClosestContactsTo`:
- before: ~3500 ms
- after: ~1500 ms

`getClosestNeighborsTo`:
- before: ~580 ms
- after: ~280 ms